Add files via upload
Hey dudes, I fixed it up a bit, deleted a bit and we are left with about 20 cards. If I missed some components feel free to add to it, there's a few that still need a couple notes and acceptance criteria but It's mostly filled in and roughly in sequential order now and starting to take shape!